# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: learnKK (런크크) API
  version: 0.1.0-bolt1
  description: >
    Bolt 1 walking skeleton contract. Covers auth, current-user profile, meetings, and admin
    approval. Errors use a uniform ErrorPayload body. Enum values are stored as varchar + CHECK
    in the DB (contract #3).
servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: auth
  - name: users
  - name: meetings
  - name: admin

paths:
  /api/auth/signup:
    post:
      tags: [auth]
      summary: Register a new user (MENTOR or MENTEE)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/auth/login:
    post:
      tags: [auth]
      summary: Log in and obtain a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/auth/logout:
    post:
      tags: [auth]
      summary: Revoke the current session
      security:
        - bearerAuth: []
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/users/me/profile:
    get:
      tags: [users]
      summary: Get the current user's profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [users]
      summary: Update the current user's profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/meetings:
    post:
      tags: [meetings]
      summary: Create a meeting (mentor only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeetingCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags: [meetings]
      summary: List recruiting meetings
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            default: recruiting
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageMeetingSummary'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/meetings/{id}:
    get:
      tags: [meetings]
      summary: Get a meeting by id
      parameters:
        - $ref: '#/components/parameters/MeetingId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetingResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/meetings/{id}/questions:
    get:
      tags: [meetings]
      summary: Get a meeting's survey questions
      parameters:
        - $ref: '#/components/parameters/MeetingId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SurveyQuestionDto'
    put:
      tags: [meetings]
      summary: Replace a meeting's survey questions (owning mentor only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MeetingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/SurveyQuestionDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SurveyQuestionDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/admin/meetings/{id}/approve:
    post:
      tags: [admin]
      summary: Approve meeting creation (T1, admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MeetingId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetingResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/admin/meetings/{id}/reject:
    post:
      tags: [admin]
      summary: Reject meeting creation (T2, admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MeetingId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetingResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Opaque server-side session token (Authorization: Bearer <token>)"

  parameters:
    MeetingId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64

  responses:
    BadRequest:
      description: Validation or business-rule failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorPayload'
    Unauthorized:
      description: Missing / invalid / expired session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorPayload'
    Forbidden:
      description: Authenticated but not authorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorPayload'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorPayload'
    Conflict:
      description: State conflict or uniqueness violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorPayload'

  schemas:
    Role:
      type: string
      enum: [MENTOR, MENTEE, ADMIN]

    MeetingStatus:
      type: string
      enum:
        - PENDING_APPROVAL
        - RECRUITING
        - READY_TO_START
        - IN_PROGRESS
        - COMPLETED
        - REJECTED
        - CANCELLED

    ErrorPayload:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: UPPER_SNAKE domain code (<DOMAIN>_<REASON>)
        message:
          type: string
          description: Human-readable Korean message
        details:
          type: object
          nullable: true
          additionalProperties: true

    SignupRequest:
      type: object
      required: [nickname, password, employeeNo, role]
      properties:
        nickname:
          type: string
        password:
          type: string
          format: password
        employeeNo:
          type: string
        role:
          type: string
          description: MENTOR or MENTEE (ADMIN is rejected)
          enum: [MENTOR, MENTEE]

    LoginRequest:
      type: object
      required: [nickname, password]
      properties:
        nickname:
          type: string
        password:
          type: string
          format: password

    SessionResponse:
      type: object
      required: [token, role]
      properties:
        token:
          type: string
        role:
          $ref: '#/components/schemas/Role'

    UserResponse:
      type: object
      required: [id, nickname, employeeNo, role]
      properties:
        id:
          type: integer
          format: int64
        nickname:
          type: string
        employeeNo:
          type: string
        role:
          $ref: '#/components/schemas/Role'

    ProfileResponse:
      type: object
      required: [nickname, employeeNo, tags]
      properties:
        nickname:
          type: string
        employeeNo:
          type: string
        tags:
          type: array
          items:
            type: string
        intro:
          type: string
          nullable: true

    ProfileUpdateRequest:
      type: object
      properties:
        tags:
          type: array
          items:
            type: string
        intro:
          type: string
          nullable: true

    MeetingCreateRequest:
      type: object
      required: [title, weeks, capacity]
      properties:
        title:
          type: string
        topic:
          type: string
          nullable: true
        weeks:
          type: integer
        recruitStart:
          type: string
          format: date-time
          nullable: true
        recruitEnd:
          type: string
          format: date-time
          nullable: true
        capacity:
          type: integer
        format:
          type: string
          nullable: true
        initialContent:
          type: string
          nullable: true

    MeetingResponse:
      type: object
      required: [id, mentorId, title, weeks, capacity, status]
      properties:
        id:
          type: integer
          format: int64
        mentorId:
          type: integer
          format: int64
        title:
          type: string
        topic:
          type: string
          nullable: true
        weeks:
          type: integer
        recruitStart:
          type: string
          format: date-time
          nullable: true
        recruitEnd:
          type: string
          format: date-time
          nullable: true
        capacity:
          type: integer
        format:
          type: string
          nullable: true
        initialContent:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/MeetingStatus'
        rejectReason:
          type: string
          nullable: true

    MeetingSummary:
      type: object
      required: [id, title, weeks, capacity, status]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        topic:
          type: string
          nullable: true
        weeks:
          type: integer
        capacity:
          type: integer
        status:
          $ref: '#/components/schemas/MeetingStatus'

    SurveyQuestionDto:
      type: object
      required: [orderNo, text, type]
      properties:
        orderNo:
          type: integer
        text:
          type: string
        type:
          type: string
        options:
          type: array
          items:
            type: string
        required:
          type: boolean

    RejectRequest:
      type: object
      properties:
        reason:
          type: string
          nullable: true

    PageMeetingSummary:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/MeetingSummary'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer